backend/services/rag-service/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
sqlalchemy
sentence-transformers
faiss-cpu
python-dotenv
httpx
